huge update to installation script
A large amount of changes are made to this script. A summary of the
changes would be:

 * couchdb is compiled without the --prefix setting
 * starts by changing into the working directory
 * install cURL (if not present)
 * download couchdb source using cURL and tar'ing the output
 * syntactic sugar using pushd/popd in favour of cd
 * use the default upstart script given by apache

The --prefix setting is great if we want to have multiple versions
of couchdb running on the same server, but we aren't doing that for
now. The default install paths are easier to remember (e.g. admins
will be expecting the logs to reside in /var/log/couchdb).

Changing directories to the working dir (scripts/, in this case)
makes it a lot easier to debug relative directory errors, and helps
us to contain everything to one directory

Using cURL to download the bits shaves off a couple LOC, so it
helps keep the script clean and tidy. We should use it since it's
already available in the Stackato VM. Attempting to install cURL
helps us port this service over to other systems for testing
purposes.

pushd and popd are great for maintaining the stack when writing
BASH scripts, so it's handy to use rather than having to remember
what directory you're currently working in, and makes the script
more maintainable.